fix(storybook-angular): use preset.mjs for Windows support
