+++ packages/storybook-angular/package.json
@@ -28,7 +28,7 @@
       "import": "./src/index.js",
       "require": "./src/index.js"
     },
-    "./preset": "./src/lib/preset.js",
+    "./preset": "./preset.mjs",
     "./package.json": "./package.json",
     "./*": "./src/*",
     "./testing": {
