+++ packages/storybook-angular/preset/preset.d.ts
@@ -1,0 +1,10 @@
+export declare const previewAnnotations: (
+  entries: any[],
+  options: any,
+) => Promise<any[]>;
+export declare const core: (config: any, options: any) => Promise<any>;
+export declare const viteFinal: (
+  config: any,
+  options: any,
+) => Promise<Record<string, any>>;
+export { addons } from '@storybook/angular/preset';
